Move test-only func from internal/internal to encoding/internal
